Cleanup

+++ .pgs/vortex/bin/clean.js
@@ -4,8 +4,23 @@ const FS = require("fs");
 const EXEC = require("child_process").exec;
 
 
+const VERBOSE = /(-\w*v|-\w*d)/.test(process.argv[2] || "");
+
+
 // TODO: Based on ignore rules of each program, remove all temporary assets
 //       for each nested in-tree program (skip symlinked programs).
+
+function getCleanIgnoreRule () {
+	var path = PATH.join(process.cwd(), ".cleanignore");
+	if (!FS.existsSync(path)) {
+		return {};
+	}
+	return FS.readFileSync(path, "utf8").split("\n").map(function(rule) {
+		return rule.replace(/\s/g, "");
+	});
+}
+
+var cleanIgnoreRules = getCleanIgnoreRule();
 
 
 var commands = [];
@@ -19,6 +34,7 @@ FS.readFileSync(PATH.join(process.cwd(), ".gitignore"), "utf8").split("\n").forE
 		return;
 	}
 	if (/^#/.test(line)) return;
+	if (cleanIgnoreRules.indexOf(line) > -1) return;
 	if (/^!\//.test(line)) {
 		commands.push('rm -Rf ' + line.substring(2));
 		commands.push("git checkout HEAD -- " + line.substring(2));
@@ -37,13 +53,23 @@ FS.readFileSync(PATH.join(process.cwd(), ".gitignore"), "utf8").split("\n").forE
 
 
 var cwd = PATH.dirname(__dirname);
-process.stdout.write("Cleaning for directory '" + cwd + "':\n");
-// TODO: Only print output on debug.
-commands.forEach(function (command) {
-	process.stdout.write(command + "\n");
-});
 
+if (VERBOSE) {
+	process.stdout.write("Cleaning for directory '" + cwd + "':\n");
+	commands.forEach(function (command) {
+		process.stdout.write(command + "\n");
+	});
+}
 
 EXEC(commands.join("; "), {
 	cwd: cwd
 });
+
+
+if (
+	FS.existsSync(PATH.join(cwd, "PINF.json")) &&
+	FS.readFileSync(PATH.join(cwd, "PINF.json"), "utf8") === "{}"
+) {
+	FS.unlinkSync(PATH.join(cwd, "PINF.json"));
+}
+
